feat(text input): enable mouse focus and selected text deletion w/ Bksp
Mouse support has been enabled for all `TextArea` widgets, to select them using the mouse pointer. This allows users to directly click on a text area to give it focus.

A new key binding for the `backspace` key has been added. This binding checks if text is currently selected within a buffer. If a selection exists, pressing `backspace` will now delete the entire selected block of text. If there is no selection, it performs the standard single-character deletion.